+++ changeLog.md
@@ -19,3 +19,4 @@ All notable changes to this project will be documented in this file.
 -Task<IEnumerable<MovieModel>> GetList(string genre); interface was added into IMovieGrainClient interface
 -Task<IEnumerable<MovieModel>> GetRatedFilms(); interface was added into IMovieGrainClient interface
 -Task Update(string key, string name); interface was added into IMovieGrainClient interface
+-MovieGrainClientTest unit tests were defined into Movies.Server.Test project but not filled yet